backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

litellm>=1.51.0
langchain-core>=0.3.26
langchain-text-splitters>=0.3.4
sentence-transformers>=3.3.0

pinecone>=5.3.0
qdrant-client>=1.12.0

python-jose[cryptography]>=3.3.0
PyJWT>=2.9.0

python-multipart>=0.0.17
aiohttp>=3.11.0
boto3>=1.35.0
pandas>=2.2.0

pytest>=8.3.0
pytest-asyncio>=0.24.0
ruff>=0.8.0
mypy>=1.14.0
httpx>=0.27.0
